+++ src/main/kotlin/com/github/frierenzk/Muppet.kt
@@ -6,6 +6,7 @@ import com.github.frierenzk.input.InputListener
 import com.github.frierenzk.server.Linkage
 import com.github.frierenzk.task.TaskPoolManager
 import com.github.frierenzk.ticker.TaskTicker
+import com.github.frierenzk.utils.MEvent
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
 import kotlin.system.exitProcess
